requirements.txt
pytube3
langchain
click
openai-whisper
transformers
nltk
PyYAML
yt-dlp